cleanup unused imports from manage cloud backups file

+++ src/hooks/useManageCloudBackups.ts
@@ -3,19 +3,12 @@ import lang from 'i18n-js';
 import { useDispatch } from 'react-redux';
 import { cloudPlatform } from '../utils/platform';
 import { WrappedAlert as Alert } from '@/helpers/alert';
-import {
-  deleteAllBackups,
-  fetchAllBackups,
-  fetchUserDataFromCloud,
-} from '@/handlers/cloudBackup';
-import { useNavigation } from '@/navigation/Navigation';
+import { deleteAllBackups } from '@/handlers/cloudBackup';
 import { clearAllWalletsBackupStatus } from '@/redux/wallets';
-import Routes from '@/navigation/routesNames';
 import { showActionSheetWithOptions } from '@/utils';
 
 export default function useManageCloudBackups() {
   const dispatch = useDispatch();
-  const { navigate } = useNavigation();
 
   const manageCloudBackups = useCallback(() => {
     const buttons = [`Delete All ${cloudPlatform} Backups`, 'Cancel'];
@@ -28,23 +21,25 @@ export default function useManageCloudBackups() {
         title: `Manage ${cloudPlatform} Backups`,
       },
       async (_buttonIndex: number) => {
-        showActionSheetWithOptions(
-          {
-            cancelButtonIndex: 1,
-            destructiveButtonIndex: 0,
-            message: `Are you sure you want to delete your ${cloudPlatform} wallet backups?`,
-            options: [`Confirm and Delete Backups`, 'Cancel'],
-          },
-          async (buttonIndex: any) => {
-            if (buttonIndex === 0) {
-              await dispatch(clearAllWalletsBackupStatus());
-              // Delete all backups (debugging)
-              await deleteAllBackups();
+        if (_buttonIndex === 0) {
+          showActionSheetWithOptions(
+            {
+              cancelButtonIndex: 1,
+              destructiveButtonIndex: 0,
+              message: `Are you sure you want to delete your ${cloudPlatform} wallet backups?`,
+              options: [`Confirm and Delete Backups`, 'Cancel'],
+            },
+            async (buttonIndex: any) => {
+              if (buttonIndex === 0) {
+                await dispatch(clearAllWalletsBackupStatus());
+                // Delete all backups (debugging)
+                await deleteAllBackups();
 
-              Alert.alert(lang.t('back_up.backup_deleted_successfully'));
+                Alert.alert(lang.t('back_up.backup_deleted_successfully'));
+              }
             }
-          }
-        );
+          );
+        }
       }
     );
   }, [dispatch]);
